LayoutTests/http/tests/security/contentSecurityPolicy/script-src-strict-dynamic-inline-report-expected.txt

CONSOLE MESSAGE: Refused to execute a script because it does not appear in the script-src directive of the Content Security Policy.
CSP report received:
CONTENT_TYPE: application/csp-report
HTTP_HOST: 127.0.0.1:8000
HTTP_REFERER: http://127.0.0.1:8000/security/contentSecurityPolicy/script-src-strict-dynamic-inline-report.py
REQUEST_METHOD: POST
REQUEST_URI: /security/contentSecurityPolicy/resources/save-report.py
=== POST DATA ===
{"type":"csp-violation","url":"http://127.0.0.1:8000/security/contentSecurityPolicy/script-src-strict-dynamic-inline-report.py","csp-report":{"document-uri":"http://127.0.0.1:8000/security/contentSecurityPolicy/script-src-strict-dynamic-inline-report.py","referrer":"","violated-directive":"script-src-elem","effective-directive":"script-src-elem","original-policy":"script-src 'nonce-dummy' 'strict-dynamic'; report-uri resources/save-report.py","blocked-uri":"inline","status-code":200}}
